--- website/src/components/ImageUpload.jsx
@@ -1,80 +1,0 @@
-import {useState} from "react";
-
-const ImageUpload = ({onImageUpload}) => {
-  const [dragActive, setDragActive] = useState(false);
-
-  const handleDrag = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    if (e.type === "dragenter" || e.type === "dragover") {
-      setDragActive(true);
-    } else if (e.type === "dragleave") {
-      setDragActive(false);
-    }
-  };
-
-  const handleDrop = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setDragActive(false);
-
-    if (e.dataTransfer.files && e.dataTransfer.files[0]) {
-      handleFile(e.dataTransfer.files[0]);
-    }
-  };
-
-  const handleChange = (e) => {
-    e.preventDefault();
-    if (e.target.files && e.target.files[0]) {
-      handleFile(e.target.files[0]);
-    }
-  };
-
-  const handleFile = (file) => {
-    if (file.type.startsWith("image/")) {
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        onImageUpload(reader.result);
-      };
-      reader.readAsDataURL(file);
-    }
-  };
-
-  return (
-    <div
-      className={`relative border-2 border-dashed rounded-lg p-8 text-center transition-colors ${
-        dragActive
-          ? "border-[#1f6feb] bg-[#1f6feb]/10"
-          : "border-[#30363d] hover:border-[#484f58]"
-      }`}
-      onDragEnter={handleDrag}
-      onDragLeave={handleDrag}
-      onDragOver={handleDrag}
-      onDrop={handleDrop}
-    >
-      <input type="file" id="file-upload" className="hidden" accept="image/*" onChange={handleChange} />
-
-      <label htmlFor="file-upload" className="cursor-pointer flex flex-col items-center justify-center">
-        <svg
-          className="w-12 h-12 mb-4 text-[#8b949e]"
-          stroke="currentColor"
-          fill="none"
-          viewBox="0 0 48 48"
-          aria-hidden="true"
-        >
-          <path
-            d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
-        <p className="mb-2 text-sm text-[#c9d1d9]">
-          <span className="font-semibold">Click to upload</span> or drag and drop
-        </p>
-      </label>
-    </div>
-  );
-};
-
-export default ImageUpload;
